can choose wavelet type. wav file example

+++ sinExample.py
@@ -10,18 +10,14 @@ cwtWidth = 256
 signal = np.sin(np.arange(sampleSize) / 20.)
 
 # Create tensorflow operations
-cwtOp = cwtMortlet(tf.float16, signal, cwtWidth)
-waveletOp = mortletWavelet(tf.float16, 32, sampleSize) # Scale of 32. We are using from 1 to 256 (cwtWidth parameter). After scale of 160 it gets less accurate with sampleSize of 1000
+cwtOp = cwtMortlet(tf.float32, signal, cwtWidth)
+waveletOp = mortletWavelet(tf.float32, 32, sampleSize) # Scale value 32. We are using from 1 to 256 (cwtWidth parameter). After scale of 160 it gets less accurate with sampleSize of 1000
 
 # Run tensorflow
 sess = tf.Session()
 cwt = sess.run(cwtOp)
 wavelet = sess.run(waveletOp)
 sess.close()
-
-# Make matplotlib happy. Does not take numpy float16
-cwt = cwt.astype(float)
-wavelet = wavelet.astype(float)
 
 # Plot signal, wavelet, cwt
 f, axarr = plt.subplots(3, sharex=True)
